tests/wpt/tests/html/dom/documents/resource-metadata-management/document-compatmode-03.html

<title>document.compatMode: Quirks</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-compatmode">
<body>
<div id="log"></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_equals(document.compatMode, "BackCompat");
})
</script>
